# openapi.yaml
openapi: 3.1.0
info:
  title: Retro Casino API
  version: 0.1.0
  description: |
    Play-money arcade casino. The server decides every outcome; the client
    sends a bet and receives a result. All credit values are int64 and are
    never floats anywhere in the system.

    Authentication uses an opaque session token in an httpOnly, SameSite=Lax,
    Secure cookie (`retro_session`). The API is only reachable through the
    Next.js BFF in production; no public CORS.

    WebSocket messages (one connection per player, multiplexed across rooms)
    are documented informally at the bottom of this file and are not part of
    the generated TS client contract.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/users/{id}/avatar:
    get:
      operationId: getUserAvatar
      tags: [user]
      summary: Serve a player's uploaded avatar image
      description: >-
        404 unless the player uploaded an image and has no preset selected.
        Responses are immutable; the version query cache-busts.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: v
          in: query
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Avatar image
          content:
            image/png:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/Error" }
...
  /api/v1/admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List or search users (moderator+)
      parameters:
        - name: query
          in: query
          schema: { type: string }
          description: Substring match on display name or email; empty lists newest first
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
      responses:
        "200":
          description: User rows
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserList" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: retro_session
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    AdminUserRow:
      type: object
      required: [id, displayName, isGuest, role, status, createdAt]
      properties:
        id: { type: integer, format: int64 }
        displayName: { type: string }
        isGuest: { type: boolean }
        email: { type: string, nullable: true }
        emailVerified: { type: boolean }
        role: { type: string }
        status: { type: string, enum: [active, banned, self_excluded] }
        statusUntil: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    AdminUserList:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items: { $ref: "#/components/schemas/AdminUserRow" }
...